refactor(toolcontext): extract shared tool session context and bind to run context
Move ToolSessionContext, Bind, ValidateRuntimeGuard, and Merge out of
internal/mcp into protocol-agnostic internal/toolcontext. Bind now uses
the owning run's context as the value authority (tenant, identity,
trace) while transport callback contexts only contribute cancellation,
so tool approvals and elicitations no longer run on value-less callback
contexts. Merge gains the missing SupportsFileInput sticky bit.
